Added (non working) create form to the Home page

+++ MijnMigraine.Web/MijnMigraine.Web.Client/Contracts/MigraineEntryDto.cs
@@ -1,3 +1,3 @@
 ﻿namespace MijnMigraine.Web.Client.Contracts;
 
-public record MigraineEntryDto(DateTime DateOfOccurrence, byte Severity, decimal Duration, string? AdditionalInfo);
+public record MigraineEntryDto(DateTime DateOfOccurrence, int Severity, decimal Duration, string? AdditionalInfo);
